--- monoruby/src/compiler/jitgen/binary_op/float_binop.rs
@@ -1,213 +1,0 @@
-use crate::bytecodegen::BinOpK;
-
-use super::*;
-
-impl Codegen {
-    pub(in crate::compiler::jitgen) fn gen_float_binop(
-        &mut self,
-        kind: BinOpK,
-        using_xmm: UsingXmm,
-        dst: Xmm,
-        mode: FMode,
-    ) {
-        match mode {
-            FMode::RR(l, r) => self.binop_float_rr(kind, using_xmm, dst, l, r),
-            FMode::RI(l, r) => self.binop_float_ri(kind, using_xmm, dst, l, r),
-            FMode::IR(l, r) => self.binop_float_ir(kind, using_xmm, dst, l, r),
-        }
-    }
-
-    fn binop_float_rr(&mut self, kind: BinOpK, using_xmm: UsingXmm, dst: Xmm, l: Xmm, r: Xmm) {
-        let lhs = l.enc();
-        let rhs = r.enc();
-        let ret = dst.enc();
-        match kind {
-            BinOpK::Add => {
-                if ret == rhs {
-                    monoasm!( &mut self.jit,
-                        addsd xmm(ret), xmm(lhs);
-                    );
-                } else {
-                    self.xmm_mov(l, dst);
-                    monoasm!( &mut self.jit,
-                        addsd xmm(ret), xmm(rhs);
-                    );
-                }
-            }
-            BinOpK::Sub => {
-                if ret == rhs {
-                    monoasm!( &mut self.jit,
-                        movq  xmm0, xmm(lhs);
-                        subsd xmm0, xmm(ret);
-                        movq  xmm(ret), xmm0;
-                    );
-                } else {
-                    self.xmm_mov(l, dst);
-                    monoasm!( &mut self.jit,
-                        subsd xmm(ret), xmm(rhs);
-                    );
-                }
-            }
-            BinOpK::Mul => {
-                if ret == rhs {
-                    monoasm!( &mut self.jit,
-                        mulsd xmm(ret), xmm(lhs);
-                    );
-                } else {
-                    self.xmm_mov(l, dst);
-                    monoasm!( &mut self.jit,
-                        mulsd xmm(ret), xmm(rhs);
-                    );
-                }
-            }
-            BinOpK::Div => {
-                if ret == rhs {
-                    monoasm!( &mut self.jit,
-                        movq  xmm0, xmm(lhs);
-                        divsd xmm0, xmm(ret);
-                        movq  xmm(ret), xmm0;
-                    );
-                } else {
-                    self.xmm_mov(l, dst);
-                    monoasm!( &mut self.jit,
-                        divsd xmm(ret), xmm(rhs);
-                    );
-                }
-            }
-            BinOpK::Exp => {
-                self.xmm_save(using_xmm);
-                monoasm!( &mut self.jit,
-                    movq xmm0, xmm(lhs);
-                    movq xmm1, xmm(rhs);
-                    movq rax, (pow_ff_f as u64);
-                    call rax;
-                );
-                self.xmm_restore(using_xmm);
-                monoasm!( &mut self.jit,
-                    movq xmm(ret), xmm0;
-                );
-            }
-            _ => unimplemented!(),
-        }
-    }
-
-    fn binop_float_ri(&mut self, kind: BinOpK, using_xmm: UsingXmm, dst: Xmm, l: Xmm, r: i16) {
-        let rhs_label = self.jit.const_f64(r as f64);
-        let ret = dst.enc();
-        match kind {
-            BinOpK::Add => {
-                self.xmm_mov(l, dst);
-                monoasm!( &mut self.jit,
-                    addsd xmm(ret), [rip + rhs_label];
-                );
-            }
-            BinOpK::Sub => {
-                self.xmm_mov(l, dst);
-                monoasm!( &mut self.jit,
-                    subsd xmm(ret), [rip + rhs_label];
-                );
-            }
-            BinOpK::Mul => {
-                self.xmm_mov(l, dst);
-                monoasm!( &mut self.jit,
-                    mulsd xmm(ret), [rip + rhs_label];
-                );
-            }
-            BinOpK::Div => {
-                self.xmm_mov(l, dst);
-                monoasm!( &mut self.jit,
-                    divsd xmm(ret), [rip + rhs_label];
-                )
-            }
-            BinOpK::Exp => {
-                let lhs = l.enc();
-                self.xmm_save(using_xmm);
-                monoasm!( &mut self.jit,
-                    movq xmm0, xmm(lhs);
-                    movq xmm1, [rip + rhs_label];
-                    movq rax, (pow_ff_f as u64);
-                    call rax;
-                );
-                self.xmm_restore(using_xmm);
-                monoasm!( &mut self.jit,
-                    movq xmm(ret), xmm0;
-                );
-            }
-            _ => unimplemented!(),
-        }
-    }
-
-    fn binop_float_ir(&mut self, kind: BinOpK, using_xmm: UsingXmm, dst: Xmm, l: i16, r: Xmm) {
-        let lhs = self.jit.const_f64(l as f64);
-        let rhs = r.enc();
-        let ret = dst.enc();
-        match kind {
-            BinOpK::Add => {
-                if ret != rhs {
-                    monoasm!( &mut self.jit,
-                        movq  xmm(ret), [rip + lhs];
-                        addsd xmm(ret), xmm(rhs);
-                    );
-                } else {
-                    monoasm!( &mut self.jit,
-                        addsd xmm(ret), [rip + lhs];
-                    );
-                }
-            }
-            BinOpK::Sub => {
-                if ret != rhs {
-                    monoasm!( &mut self.jit,
-                        movq  xmm(ret), [rip + lhs];
-                        subsd xmm(ret), xmm(rhs);
-                    );
-                } else {
-                    monoasm!( &mut self.jit,
-                        movq  xmm0, xmm(rhs);
-                        movq  xmm(ret), [rip + lhs];
-                        subsd xmm(ret), xmm0;
-                    );
-                }
-            }
-            BinOpK::Mul => {
-                if ret != rhs {
-                    monoasm!( &mut self.jit,
-                        movq  xmm(ret), [rip + lhs];
-                        mulsd xmm(ret), xmm(rhs);
-                    );
-                } else {
-                    monoasm!( &mut self.jit,
-                        mulsd xmm(ret), [rip + lhs];
-                    );
-                }
-            }
-            BinOpK::Div => {
-                if ret != rhs {
-                    monoasm!( &mut self.jit,
-                        movq  xmm(ret), [rip + lhs];
-                        divsd xmm(ret), xmm(rhs);
-                    );
-                } else {
-                    monoasm!( &mut self.jit,
-                        movq  xmm0, xmm(ret);
-                        movq  xmm(ret), [rip + lhs];
-                        divsd xmm(ret), xmm0;
-                    );
-                }
-            }
-            BinOpK::Exp => {
-                self.xmm_save(using_xmm);
-                monoasm!( &mut self.jit,
-                    movq xmm0, [rip + lhs];
-                    movq xmm1, xmm(rhs);
-                    movq rax, (pow_ff_f as u64);
-                    call rax;
-                );
-                self.xmm_restore(using_xmm);
-                monoasm!( &mut self.jit,
-                    movq xmm(ret), xmm0;
-                );
-            }
-            _ => unimplemented!(),
-        }
-    }
-}
